requirements.txt
mutagen==1.38
boto3==1.4.5
click==6.7
unicode_slugify==0.1.3
PyYAML==3.12
